Repository e Service/ Implementando o FactoryMethod

+++ src/main/java/com/cauandev/estoqueremedio/model/TipoRemedio.java
@@ -4,6 +4,8 @@ import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.MappedSuperclass;
+import org.springframework.format.annotation.DateTimeFormat;
+import org.springframework.format.datetime.DateFormatter;
 
 import java.time.LocalDate;
 
@@ -17,11 +19,11 @@ public abstract class TipoRemedio {
     private Tarja tarja;
     private Formula formula;
 
-    public int getCode() {
+    public Long getCode() {
         return code;
     }
 
-    public void setCode(int code) {
+    public void setCode(Long code) {
         this.code = code;
     }
